Cleared up the scripts duplications
