+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
-exclude = .git, __pycache__, docs, Stock LSTM
-ignore = E501
+exclude = .git, __pycache__, mlruns, docs, LSTM_predictor
+ignore = D100, D103, E501
