+++ Flask_webapp/templates/index.html
@@ -1,0 +1,106 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta http-equiv="X-UA-Compatible" content="ie=edge">
+<!--  -->
+<style>
+	body, html {
+	    margin: 0;
+	    font-family: sans-serif;
+	    background-color: lightblue;
+	}
+
+	.content {
+	    margin:  auto;
+	    width: 400px;
+	}
+
+	div {
+	  	border: 0px solid black;
+	  	margin: 25px 50px 75px 100px;
+	  	background-color: lightblue;
+	}
+
+	table, td, th {
+	    border: 1px solid #aaa;
+	}
+
+	table {
+	    border-collapse: collapse;
+	    width: 50%;
+	}
+	table.center {
+			margin-left: auto;
+			margin-right: auto;
+	}
+
+	th {
+	    height: 30px;
+	}
+
+	td {
+	    text-align: center;
+	    padding: 5px;
+	}
+
+	.form {
+	    margin-top: 20px;
+	}
+
+  /* input[type='file'] {color: red;} */
+
+	#content {
+	    width: 50%;
+	}
+  hr {width:100%; text-align:left;margin-left:0;}
+</style>
+<!--  -->
+<title>Stock Prediction</title>
+</head>
+
+<!--  -->
+<!--  -->
+
+<body>
+<div>
+
+<!-- action='/predict' sends the form information to /predict endpoint, or page -->
+<form action="/predict" method="post" enctype="multipart/form-data">
+<table class="center">
+	<tr>
+    <td>Ticker Name </td>
+		<td>Select Model</td>
+	</tr>
+<!--  -->
+  <tr>
+    <!-- added ticker name textbox -->
+    <td>
+    <input type = 'text' name = 'ticker'/>
+    </td>
+    <td>
+      <select name="comp_select" class="selectpicker form-control">
+      {% for m in modelfiles %}
+      <option value="{{ m }}">{{ m }}</option>
+      {% endfor %}
+      </select>
+    </td>
+  </tr>
+    <td></td>
+  <tr>
+
+  </tr>
+</table>
+<button type="submit" onclick="">Predict</button>
+</form>
+
+<hr>
+How to use this stock predictor:<br><br>
+<li>Provide a valid stock ticker.</li>
+<li>Select a model. </li>
+
+</div>
+</body>
+</html>
