+++ LSTM_predictor/setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup
+# name specifies the folder for the actual .py files
+setup(name="stockPredictor", version="0.2.0")
