--- scripts/stockPredictor/requirements.txt
@@ -1,7 +1,0 @@
-tensorflow==2.0.1
-Keras==2.3.1
-numpy==1.18.5
-pandas==1.0.4
-matplotlib==3.2.1
-yfinance==0.1.54
-scikit_learn==0.23.1
