--- scripts/train&predict/requirements.txt
@@ -1,7 +1,0 @@
-numpy==1.18.5
-Keras==2.3.1
-pandas==1.0.4
-matplotlib==3.2.1
-tensorflow==2.0.1
-scikit_learn==0.23.1
-yfinance==0.1.54
