--- webapp/.flake8
@@ -1,3 +1,0 @@
-[flake8]
-exclude = .git, __pycache__, mlruns
-ignore = D100, D103, E501
