--- webapp/.gitignore
@@ -1,2 +1,0 @@
-__pycache__
-mlruns
