--- webapp/mypackage/.gitignore
@@ -1,2 +1,0 @@
-__pycache__
-*.egg-info
