--- webapp/mypackage/setup.py
@@ -1,3 +1,0 @@
-from setuptools import setup
-# name specifies the folder for the actual .py files
-setup(name="script", version="0.1.0")
