--- webapp/readme.md
@@ -1,22 +1,0 @@
-# Web-App for LSTM powered stock prediction
-
-## Development
-Set up environment
-
-* Setup up virtual environment with Anaconda, or VENV
-
->
-    $ conda create --name env_name
->
-    $ conda activate env_name
-
-* Install Python 3.7.6
-
-* Install dependencies with pip:
-
->
-    $ pip install -r requirements.txt
-
-## Dataset
-Example datasets are given [here] (https://github.com/Ling-Jun/LSTM-Stock-price/tree/master/time%20series%20data).
-* We now allow users to provide a simple stock ticker, the price data will be fetched from Yahoo finance.
