--- webapp/templates/predict.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-<!--  -->
-		<style>
-			body, html {
-			    margin: 0;
-			    font-family: sans-serif;
-			    background-color: lightblue;
-			}
-
-
-			div {
-			  	border: 0px solid black;
-			  	margin: 50px 50px 50px 50px;
-			  	background-color: lightblue;
-          width: 50%;
-			}
-
-		</style>
-<!--  -->
-		<title>Addition</title>
-</head>
-
-<!--  -->
-<!--  -->
-
-<body>
-	<div>
-  <form action="/" method="get, post">
-  <!--Form Action attribute specifies where to send the form-data when a form is submitted -->
-	<button type="submit">Back</button>
-	</form>
-  
-  <br>
-  <b>RMSE</b> is: {{rmse}}
-  <br>
-  Model is: {{model}}
-  <br>
-  <img src="data:image/png;base64, {{ img }}">
-  </div>
-</body>
-</html>
