--- webapp/test/.gitignore
@@ -1,1 +1,0 @@
-__pycache__, mlruns
