--- webapp/test/templates/temp.html
@@ -1,6 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<body>
-{{name}}
-</body>
-</html>
